resolvido metade do desafio 1
